compose the request logger in the original logger itself
